requirements.txt
Gunicorn==20.0.4
Flask==2.1.1
mysql-connector-python